feat: move states to their own packages
